be/requirements.txt
alembic==1.11.2
annotated-types==0.7.0
anyio==3.6.1
attrs==20.3.0
bcrypt==3.2.0
click==7.1.2
dnspython==2.1.0
fastapi==0.115.5
FastAPI-SQLAlchemy==0.2.1
greenlet==3.1.1
h11==0.12.0
idna==2.10
iniconfig==1.1.1
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
passlib==1.7.4
psycopg2==2.9.10
py==1.10.0
pyasn1==0.6.1
pycparser==2.20
pydantic==2.10.2
pydantic_core==2.27.1
pydantic-settings==2.6.1
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==1.0.1
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.4
requests==2.25.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.41.3
typing_extensions==4.12.2
urllib3>=1.26.5
Authlib==1.3.0
uvicorn==0.22.0
redis>=5.0.0,<6.0.0
itsdangerous==2.0.1
httpx==0.20.0
faker
cloudinary==1.42.1

